src/requirements.txt
python-dotenv==1.0.1
influxdb_client==1.46.0